+++ backend/src/main/java/com/funeat/tag/dto/TagDto.java
@@ -1,4 +1,6 @@
-package com.funeat.tag.domain;
+package com.funeat.tag.dto;
+
+import com.funeat.tag.domain.Tag;
 
 public class TagDto {
